requirements.txt
requests==2.31.0
python-dotenv==1.0.0
rich==13.7.0
cryptography==42.0.0
schedule==1.2.1
